Backend and frontend refactored settings
