+++ back-end/package.json
@@ -15,43 +15,45 @@
     "lint:fix": "eslint --fix",
     "format": "prettier --config .prettierrc.json --write .",
     "start": "npm run build && node --env-file=.env.prod dist/index",
-    "build": "rm -rf ./dist/ && tsc",
+    "prebuild": "rimraf ./dist/",
+    "build": "tsc && cp -r package*.json dist/ && cp .env.example dist/.env.example",
     "migration:generate": "typeorm-ts-node-commonjs -d ./src/data-source.ts migration:generate ./src/migrations/initial",
     "migration:run": "typeorm-ts-node-commonjs -d ./src/data-source.ts migration:run"
   },
   "dependencies": {
     "bcrypt": "^5.1.1",
-    "better-sqlite3": "^11.7.0",
-    "bindings": "^1.5.0",
-    "cache-manager": "^6.1.0",
+    "better-sqlite3": "^9.0.0",
+    "cache-manager": "^6.3.2",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^5.0.1",
     "express-rate-limit": "^7.5.0",
     "jsonwebtoken": "^9.0.2",
-    "mysql2": "^3.11.5",
-    "pg": "^8.13.0",
+    "mysql2": "^3.12.0",
+    "pdfmake": "^0.3.0-beta.14",
+    "pg": "^8.13.1",
     "reflect-metadata": "^0.2.2",
     "sqlite": "^5.1.1",
     "typeorm": "^0.3.20",
     "typeorm-naming-strategies": "^4.1.0",
     "uuid": "^11.0.3",
-    "winston": "^3.15.0",
-    "zod": "^3.23.8"
+    "winston": "^3.17.0",
+    "zod": "^3.24.1"
   },
   "devDependencies": {
-    "@eslint/js": "^9.13.0",
+    "@eslint/js": "^9.17.0",
     "@types/bcrypt": "^5.0.2",
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
     "@types/jsonwebtoken": "^9.0.7",
-    "@types/node": "^22.7.7",
+    "@types/node": "^22.10.2",
+    "@types/pdfmake": "^0.2.10",
     "@types/uuid": "^10.0.0",
-    "eslint": "^9.13.0",
-    "globals": "^15.11.0",
+    "eslint": "^9.17.0",
     "prettier": "^3.4.2",
+    "rimraf": "^6.0.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.6.3",
-    "typescript-eslint": "^8.18.1"
+    "typescript-eslint": "^8.18.2"
   }
 }
