+++ front-end/src/app/services/api.service.ts
@@ -1,5 +1,5 @@
 import { HttpClient } from '@angular/common/http';
-import { environment } from '../environments/environment';
+import { environment } from '../../environments/environment';
 import { Observable } from 'rxjs';
 import { Injectable, inject } from '@angular/core';
