+++ front-end/src/app/services/auth.service.ts
@@ -2,7 +2,7 @@ import { BehaviorSubject, Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { EventEmitter, Injectable } from '@angular/core';
 
-import { environment } from '../environments/environment';
+import { environment } from '../../environments/environment';
 import { Router } from '@angular/router';
 import { OrgYear, Organization, LoginResponse } from '../models';
