+++ front-end/src/environments/environment.prod.ts
@@ -1,0 +1,5 @@
+export const environment = {
+  production: true,
+  ICMS_API_URL: 'http://localhost/api',
+  AUTH_SERVER_API_URL: 'http://localhost/',
+};
